+++ src/cloudpilot/native/util.h
@@ -11,7 +11,7 @@ namespace util {
         "PalmM500",  "PalmM505",        "PalmM515", "PalmM105",        "PalmM100",  "PalmM125",
         "PalmM130",  "PalmI705",        "PalmI710", "HandEra330",      "PEG-S300",  "PEG-S320",
         "PEG-T400",  "PEG-N600C/N610C", "PEG-T600", "PEG-N700C/N710C", "YSX1230",   "NR70",
-        "PEG-S500C", "HandEra330c"};
+        "PEG-S500C", "HandEra330c",     "PalmM520"};
 
     bool readFile(string file, unique_ptr<uint8[]>& buffer, size_t& len);