+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 1.7.10
+
+-   Fix a heap overflow.
+
 # Version 1.7.9
 
 -   Fix status bar color in light mode on iOS.